refactored

+++ src/main/java/com/bionic/edu/customer/CustomerDaoImpl.java
@@ -7,9 +7,6 @@ import org.springframework.transaction.annotation.Transactional;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
-import javax.persistence.criteria.CriteriaBuilder;
-import javax.persistence.criteria.CriteriaQuery;
-import javax.persistence.criteria.Root;
 import java.util.List;
 
 @Repository
@@ -23,29 +20,22 @@ public class CustomerDaoImpl implements CustomerDao {
     }
 
     @Override
-    public List<Customer> getAllCustomer() {
-        CriteriaBuilder cb = em.getCriteriaBuilder();
-        CriteriaQuery<Customer> cq = cb.createQuery(Customer.class);
-        Root<Customer> rootEntry = cq.from(Customer.class);
-        CriteriaQuery<Customer> all = cq.select(rootEntry);
-        TypedQuery<Customer> allQuery = em.createQuery(all);
-        return allQuery.getResultList();
+    public List<Customer> findAll() {
+        TypedQuery<Customer> tq = em.createQuery("SELECT c FROM Customer c",Customer.class);
+        return tq.getResultList();
     }
 
     @Transactional
     @Override
     public void save(Customer c) {
         em.persist(c);
-        em.flush();
     }
 
     @Transactional
     @Override
     public void remove(int id) {
         Customer c = findById(id);
-        if(c != null){
-            em.remove(c);
-        }
+        if(c != null) em.remove(c);
     }
 
     @Override
